Calculate points server-side, from the record
`updatePlayerPointsAction` took a player name and a number. It was a bare
'use server' function with no `authActionClient` wrapper, no ownership check
and no validation, so any signed-in member could set any player's points to
anything — and points are what the leaderboard ranks on. `CalculatorHero`
called it from an effect on every recalculation, which is why it existed.

Points are a function of the stored record, so they are now computed from it.
`calculatePlayerPoints` is the calculation that was already running inline in
`check-auto-rank`, lifted out unchanged, and `processPlayerData` calls it
after writing the record. That is now the only place `players.points` is
written, and it covers every path that already funnels through it: the
calculator page load, the auto-rank check, and the nightly batch refresh.

The browser still calculates a running total for display. It just no longer
gets a vote on what is stored.

If the calculation throws — it needs live wiki drop rates — the stored total
is left alone rather than zeroed. A stale number is a cosmetic lag on the
leaderboard; a zero is a visible catastrophe.

`check-auto-rank` loses ~80 lines of inline arithmetic and 20 imports, and
now calls the shared function for the prospective rank alone.

`jest.setup.ts` mocked the deleted action, which failed every suite at setup
until removed. Worth noting the mock existed because a component tree was
reaching for postgres during tests — that reason is gone with the write.

Scoped down from the `commitPlayerState` refactor in the original plan: with
the field-ownership model dropped (#85), consolidating the remaining write
paths is mostly relocation, while this half is a live security fix and
removes a whole sync point.
