+++ CLAUDE.md
@@ -139,7 +139,7 @@ The page (`[player]/rank-calculator.tsx`, shared with the readonly moderator vie
 Three things to preserve when touching this page:
 - **The `aria-label`s are the test contract.** Both the Jest specs and the Cypress e2e suite query by label (`total combat points`, `combat points remaining`, `point scaling`, …). Rename a label and you break both.
 - **Category inputs live in modals, so tests must open them first.** Jest specs click the tile and await the dialog; the Cypress helper `generateScalingTests` takes an `openCategory` regex for the same reason. Totals stay on the tile, so only per-input assertions need the modal.
-- **The player-points write lives in `CalculatorHero` only.** It used to sit inside `useRank`, so every consumer of `useRankCalculator` (hero, nav bar, navigation actions) fired its own duplicate write on each recalculation. `useRank`/`useRankCalculator` are now pure and safe to call anywhere; `useTotalPoints` remains the cheaper option when only the number is needed.
+- **Nothing on the client writes points.** `useRank` / `useRankCalculator` / `useTotalPoints` are pure — the running total they produce is for display only, and `useTotalPoints` remains the cheaper option when just the number is needed. `players.points` is recalculated server-side by `calculatePlayerPoints` (`app/rank-calculator/utils/calculate-player-points.ts`) inside `processPlayerData`, from the stored record, and that is the only place it is written. Don't reintroduce a client-side write: the previous one posted to an **unauthenticated** action taking `(playerName, points)`, so the leaderboard was whatever the client last claimed, for whichever player it named. If `calculatePlayerPoints` throws (it needs live wiki drop rates) the stored total is left alone — stale is a cosmetic lag, zero would be a catastrophe.
 
 Data-source convention: return `{ success: true, data } | { success: false, error }`; filter to `players.isActive`; `isMaxed = totalLevel === 2376`; pets = `playerAcquiredItems.itemId in AllPetItemIds`; infernal = `tzhaarCape === 'Infernal cape'`.
