+++ apps/web/app/api/check-auto-rank/route.ts
@@ -9,19 +9,7 @@ import {
 } from 'discord-api-types/v10';
 import { NextRequest, NextResponse } from 'next/server';
 import { fetchPlayerDetails } from '@/app/rank-calculator/data-sources/fetch-player-details/fetch-player-details';
-import { calculateAchievementDiaryPoints } from '@/app/rank-calculator/utils/calculators/calculate-achievement-diary-points';
-import { calculateCollectionLogAndCluesPoints } from '@/app/rank-calculator/utils/calculators/calculate-collection-log-and-clues-points';
-import { calculateCollectionLogSlotPoints } from '@/app/rank-calculator/utils/calculators/calculate-collection-log-slot-points';
-import { calculateCombatAchievementPoints } from '@/app/rank-calculator/utils/calculators/calculate-combat-achievement-points';
-import { calculateCombatPoints } from '@/app/rank-calculator/utils/calculators/calculate-combat-points';
-import { calculateEhbPoints } from '@/app/rank-calculator/utils/calculators/calculate-ehb-points';
-import { calculateEhpPoints } from '@/app/rank-calculator/utils/calculators/calculate-ehp-points';
-import { calculateNotableItemsPoints } from '@/app/rank-calculator/utils/calculators/calculate-notable-items-points';
-import { calculateRank } from '@/app/rank-calculator/utils/calculators/calculate-rank';
-import { calculateScaling } from '@/app/rank-calculator/utils/calculators/calculate-scaling';
-import { calculateSkillingPoints } from '@/app/rank-calculator/utils/calculators/calculate-skilling-points';
-import { calculateTotalLevelPoints } from '@/app/rank-calculator/utils/calculators/calculate-total-level-points';
-import { calculateTotalPoints } from '@/app/rank-calculator/utils/calculators/calculate-total-points';
+import { calculatePlayerPoints } from '@/app/rank-calculator/utils/calculate-player-points';
 import { getRankName } from '@/app/rank-calculator/utils/get-rank-name';
 import { isRankUp } from '@/app/rank-calculator/utils/is-rank-up';
 import { canApplyForRank } from '@/config/ranks';
@@ -30,21 +18,7 @@ import { clientConstants } from '@/config/constants.client';
 import { rankUpMessagesKey } from '@/config/redis';
 import { discordBotClient } from '@/discord';
 import { redis } from '@/redis';
-import {
-  fetchItemDropRates,
-  generateRequiredItemList,
-} from '@/app/rank-calculator/data-sources/fetch-dropped-item-info';
-import { buildNotableItemList } from '@/app/rank-calculator/utils/build-notable-item-list';
-import { calculateAchievementDiaryCapePoints } from '@/app/rank-calculator/utils/calculators/calculate-achievement-diary-cape-points';
-import { calculateTzhaarCapePoints } from '@/app/rank-calculator/utils/calculators/calculate-tzhaar-cape-points';
-import { calculateBloodTorvaPoints } from '@/app/rank-calculator/utils/calculators/calculate-blood-torva-points';
-import { calculateDizanasQuiverPoints } from '@/app/rank-calculator/utils/calculators/calculate-dizanas-quiver-points';
-import { calculateClueScrollPoints } from '@/app/rank-calculator/utils/calculators/calculate-clue-scroll-points';
-import { calculateRadiantOathplatePoints } from '@/app/rank-calculator/utils/calculators/calculate-radiant-oathplate-points';
-import {
-  processPlayerData,
-  updatePlayerPoints,
-} from '@/lib/db/player-operations';
+import { processPlayerData } from '@/lib/db/player-operations';
 
 export async function GET(request: NextRequest) {
   try {
@@ -63,31 +37,8 @@ export async function GET(request: NextRequest) {
       throw new Error('Failed to fetch player details');
     }
 
-    const {
-      joinDate,
-      collectionLogTotal,
-      collectionLogCount,
-      acquiredItems,
-      achievementDiaries,
-      ehp,
-      ehb,
-      totalLevel,
-      combatAchievementTier,
-      currentRank,
-      hasThirdPartyData,
-      playerName,
-      accountType,
-      tzhaarCape,
-      hasBloodTorva,
-      hasRadiantOathplate,
-      hasDizanasQuiver,
-      hasAchievementDiaryCape,
-      collectionLogBonusPoints,
-      combatBonusPoints,
-      notableItemsBonusPoints,
-      skillingBonusPoints,
-      clueScrollCounts,
-    } = playerDetails.data;
+    const { currentRank, hasThirdPartyData, playerName, accountType } =
+      playerDetails.data;
 
     if (!hasThirdPartyData) {
       return NextResponse.json({ success: true });
@@ -100,91 +51,12 @@ export async function GET(request: NextRequest) {
       // Continue with rank calculation even if database sync fails
     }
 
-    const dropRates = await fetchItemDropRates([...generateRequiredItemList()]);
-    const items = Object.entries(await buildNotableItemList(dropRates));
-    const scaling = calculateScaling(joinDate);
-    const collectionLogSlotPoints = calculateCollectionLogSlotPoints(
-      collectionLogCount,
-      scaling,
-    );
-    const { totalPoints: clueScrollPoints } = calculateClueScrollPoints(
-      clueScrollCounts,
-      scaling,
-    );
-    const { pointsAwarded: totalCollectionLogPoints } =
-      calculateCollectionLogAndCluesPoints(
-        collectionLogSlotPoints,
-        collectionLogTotal,
-        clueScrollPoints,
-        collectionLogBonusPoints,
-        0.0,
-        scaling,
-      );
-    const { pointsAwarded: totalNotableItemsPoints } =
-      calculateNotableItemsPoints(
-        items,
-        acquiredItems,
-        notableItemsBonusPoints,
-        scaling,
-      );
-    const { pointsAwarded: achievementDiariesPoints } =
-      calculateAchievementDiaryPoints(achievementDiaries, scaling);
-    const ehpPoints = calculateEhpPoints(ehp, scaling);
-    const totalLevelPoints = calculateTotalLevelPoints(totalLevel, scaling);
-    const achievementDiaryCapePoints = calculateAchievementDiaryCapePoints(
-      hasAchievementDiaryCape,
-      scaling,
-    );
-    const { pointsAwarded: totalSkillingPoints } = calculateSkillingPoints(
-      achievementDiariesPoints,
-      ehpPoints,
-      totalLevelPoints,
-      achievementDiaryCapePoints,
-      skillingBonusPoints,
-      scaling,
-    );
-    const ehbPoints = calculateEhbPoints(ehb);
-    const combatAchievementTierPoints = calculateCombatAchievementPoints(
-      combatAchievementTier,
-      scaling,
-    );
-    const tzhaarCapePoints = calculateTzhaarCapePoints(tzhaarCape, scaling);
-    const bloodTorvaPoints = calculateBloodTorvaPoints(hasBloodTorva, scaling);
-    const radiantOathplatePoints = calculateRadiantOathplatePoints(
-      hasRadiantOathplate,
-      scaling,
-    );
-    const dizanasQuiverPoints = calculateDizanasQuiverPoints(
-      hasDizanasQuiver,
-      scaling,
-    );
-    const { pointsAwarded: totalCombatPoints } = calculateCombatPoints(
-      ehbPoints,
-      combatAchievementTierPoints,
-      tzhaarCapePoints,
-      bloodTorvaPoints,
-      radiantOathplatePoints,
-      dizanasQuiverPoints,
-      combatBonusPoints,
-      scaling,
-    );
-    const totalPointsAwarded = calculateTotalPoints(
-      totalCollectionLogPoints,
-      totalNotableItemsPoints,
-      totalSkillingPoints,
-      totalCombatPoints,
-    );
-    const { rank } = calculateRank(
-      acquiredItems,
-      combatAchievementTier,
-      totalPointsAwarded,
-      accountType,
-    );
-
-    // Update player points in database
-    await updatePlayerPoints(playerName, totalPointsAwarded).catch((error) => {
-      console.error(`Failed to update points for player ${playerName}:`, error);
-    });
+    // Only the prospective rank is wanted here — `processPlayerData` above has
+    // already recalculated and stored the points total. This runs the same
+    // calculation a second time rather than plumbing a return value through,
+    // which is cheap: the one expensive input, the wiki drop rates, is behind
+    // `unstable_cache`, and the rest is arithmetic.
+    const { rank } = await calculatePlayerPoints(playerDetails.data);
 
     // Same rule as the calculator's rank-up dialog: only a genuine promotion up
     // the ladder this account is scored against, and only for an account that
@@ -192,7 +64,10 @@ export async function GET(request: NextRequest) {
     // action will refuse is worse than saying nothing. A staff member's stored
     // rank is an in-game staff rank, which is on no ladder, so a bare
     // inequality nudged them about a rank they may already hold.
-    if (canApplyForRank(accountType) && isRankUp(currentRank, rank, accountType)) {
+    if (
+      canApplyForRank(accountType) &&
+      isRankUp(currentRank, rank, accountType)
+    ) {
       const hashKey = `${discordId}:${player.toLowerCase()}`;
       const previousMessageRank = await redis.hget(rankUpMessagesKey, hashKey);
